+++ DotNetLicenses.TestFramework/DisposableDirectory.fs
@@ -6,16 +6,17 @@ namespace DotNetLicenses.TestFramework
 
 open System
 open System.IO
+open TruePath
 
 type DisposableDirectory =
-    { Path: string }
+    { Path: StrictAbsolutePath }
 
     interface IDisposable with
         member this.Dispose() =
-            Directory.Delete(this.Path, true)
+            Directory.Delete(this.Path.Value, true)
 
     static member Create() =
-        let path = Path.GetTempFileName()
-        File.Delete(path)
-        Directory.CreateDirectory(path) |> ignore
+        let path = StrictAbsolutePath <| Path.GetTempFileName()
+        File.Delete path.Value
+        Directory.CreateDirectory path.Value |> ignore
         { Path = path }