+++ DotNetLicenses.TestFramework/ZipFiles.fs
@@ -6,9 +6,10 @@ module DotNetLicenses.TestFramework.ZipFiles
 
 open System.IO
 open System.IO.Compression
+open TruePath
 
-let SingleFileArchive(archivePath: string, fileRelativePath: string, fileContent: byte[]): unit =
-    use stream = File.OpenWrite archivePath
+let SingleFileArchive(archivePath: AbsolutePath, fileRelativePath: string, fileContent: byte[]): unit =
+    use stream = File.OpenWrite archivePath.Value
     use archive = new ZipArchive(stream, ZipArchiveMode.Create)
     use entry = archive.CreateEntry(fileRelativePath).Open()
     entry.Write(fileContent, 0, fileContent.Length)