+++ DotNetLicenses.Tests/SourcesTests.fs
@@ -14,30 +14,31 @@ open DotNetLicenses
 open DotNetLicenses.Sources
 open DotNetLicenses.TestFramework
 open JetBrains.Lifetimes
+open TruePath
 open Xunit
 
 [<Fact>]
 let ``Sources are enumerated for a file-system directory``(): Task = task {
     use directory = DisposableDirectory.Create()
-    do! File.WriteAllTextAsync(Path.Combine(directory.Path, "file.txt"), "content")
-    Directory.CreateDirectory(Path.Combine(directory.Path, "subdirectory")) |> ignore
-    do! File.WriteAllTextAsync(Path.Combine(directory.Path, "subdirectory", "file.txt"), "content")
+    do! File.WriteAllTextAsync((directory.Path / "file.txt").Value, "content")
+    Directory.CreateDirectory (directory.Path / "subdirectory").Value |> ignore
+    do! File.WriteAllTextAsync((directory.Path / "subdirectory" / "file.txt").Value, "content")
 
-    let spec = Directory directory.Path
+    let spec = Directory <| directory.Path.AsRelative()
     use ld = new LifetimeDefinition()
-    let! entries = ReadEntries ld.Lifetime directory.Path [| spec |]
+    let! entries = ReadEntries ld.Lifetime (AbsolutePath.op_Implicit directory.Path) [| spec |]
     Assert.Equivalent([|
-        FileSourceEntry(spec, Path.Combine(directory.Path, "file.txt"))
-        FileSourceEntry(spec, Path.Combine(directory.Path, "subdirectory", "file.txt"))
+        FileSourceEntry(directory.Path, spec, directory.Path / "file.txt")
+        FileSourceEntry(directory.Path, spec, directory.Path / "subdirectory" / "file.txt")
     |], entries)
 }
 
 [<Fact>]
 let ``SourceRelativePath is generated for a file set``(): unit =
     use directory = DisposableDirectory.Create()
-    let source = Directory <| directory.Path
-    let file = Path.Combine(directory.Path, Path.Combine("foo", "file.txt"))
-    let entry = FileSourceEntry(source, file) :> ISourceEntry
+    let source = Directory <| directory.Path.AsRelative()
+    let file = directory.Path / "foo" / "file.txt"
+    let entry = FileSourceEntry(directory.Path, source, file) :> ISourceEntry
     Assert.Equal("foo/file.txt", entry.SourceRelativePath)
 
 [<Fact>]
@@ -49,10 +50,10 @@ let ``SourceEntry calculates its hash correctly``(): Task = task {
     use directory = DisposableDirectory.Create()
     let content = "Hello"B
     let expectedHash = calculateSha256 content
-    let path = Path.Combine(directory.Path, "file.txt")
-    File.WriteAllBytes(path, content)
-    let source = Directory directory.Path
-    let entry = FileSourceEntry(source, path) :> ISourceEntry
+    let path = directory.Path / "file.txt"
+    File.WriteAllBytes(path.Value, content)
+    let source = Directory <| directory.Path.AsRelative()
+    let entry = FileSourceEntry(directory.Path, source, path) :> ISourceEntry
     let! actualHash = entry.CalculateHash()
     Assert.Equal(expectedHash, actualHash)
 }
@@ -62,11 +63,11 @@ let ``SourceEntry reads contents from a ZIP archive``(): Task = task {
     let content = "Hello"
     let fileName = "file.txt"
     use directory = DisposableDirectory.Create()
-    let archivePath = Path.Combine(directory.Path, "archive.zip")
-    ZipFiles.SingleFileArchive(archivePath, fileName, Encoding.UTF8.GetBytes content)
-    use stream = File.OpenRead archivePath
+    let archivePath = directory.Path / "archive.zip"
+    ZipFiles.SingleFileArchive(AbsolutePath.op_Implicit archivePath, fileName, Encoding.UTF8.GetBytes content)
+    use stream = File.OpenRead archivePath.Value
     use archive = new ZipArchive(stream)
-    let source = Zip archivePath
+    let source = Zip <| LocalPathPattern archivePath.Value
     let entry = ZipSourceEntry(source, archive, archivePath, fileName) :> ISourceEntry
     use! stream = entry.ReadContent()
     use reader = new StreamReader(stream)